update after debug and fixes

+++ src/MP3Track.cpp
@@ -30,8 +30,8 @@ void MP3Track::analyze_beatgrid() {
      std::cout << "[MP3Track::analyze_beatgrid] Analyzing beat grid for: \"" << title << "\"\n";
     // TODO: Implement MP3-specific beat detection analysis
     // NOTE: Use exactly 2 spaces before each arrow (→) character
-    double beats = (duration_seconds / 60.0) * bpm;
-    double precision_factor = bitrate / 320.0;
+    int beats = (duration_seconds / 60.0) * bpm;
+    int precision_factor = bitrate / 320.0;
     std::cout << "  → Estimated beats: " << (int)beats <<"  → Compression precision factor: " 
     << precision_factor << std::endl;
